+++ R/utils.R
@@ -57,7 +57,6 @@ N <- function(z) length(z)
 NAS <- function(z) sum(is.na(z))
 
 
-#' @importFrom uuid UUIDgenerate
 as_bookmark <- function(id, str) {
   new_id <- UUIDgenerate()
   bm_start_str <- sprintf("<w:bookmarkStart w:id=\"%s\" w:name=\"%s\"/>", new_id, id)
@@ -188,19 +187,4 @@ if (!"gregexec" %in% getNamespaceExports("base")) {
       res
     }
   }
-}
-.url_special_chars <- list(
-  `&` = '&amp;',
-  `<` = '&lt;',
-  `>` = '&gt;',
-  `'` = '&#39;',
-  `"` = '&quot;',
-  ` ` = "&nbsp;"
-)
-officer_url_encode <- function(x) {
-  for (chr in names(.url_special_chars)) {
-    x <- gsub(chr, .url_special_chars[[chr]], x, fixed = TRUE, useBytes = TRUE)
-  }
-  Encoding(x) <- "UTF-8"
-  x
 }